+++ src/main/java/fi/dy/masa/malilib/gui/widget/ColorIndicatorWidget.java
@@ -40,7 +40,7 @@ public class ColorIndicatorWidget extends BaseWidget
     {
         int x = this.getX();
         int y = this.getY();
-        int z = this.getZLevel();
+        float z = this.getZLevel();
         int width = this.getWidth();
         int height = this.getHeight();
